Add files via upload
There are ZERO imports in this file. 
The code could have been less number of lines, if i could have done some imports like pandas, random, etc. I have not done all of that to ensure the size of executable is small and it is more fun to work with basics than using all the APIs.